+++ opal/tests/test_patient.py
@@ -1,11 +1,8 @@
 """
 Unittests for Patients
 """
-from mock import patch
 from opal.core.test import OpalTestCase
-
 from opal.models import Patient, Episode
-from opal.tests.models import InvisibleDog
 
 
 class PatientTest(OpalTestCase):
@@ -23,7 +20,7 @@ class PatientTest(OpalTestCase):
     def test_get_active_episode(self):
         self.patient.create_episode()
         episode2 = self.patient.create_episode()
-        episode2.set_tag_names(['microbiology'], None)
+        episode2.set_tag_names(['inpatient'], None)
         self.assertEqual(episode2.id, self.patient.get_active_episode().id)
 
     def test_get_active_episode_with_no_episodes(self):